requirements.txt
openai>=1.37.0
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
streamlit>=1.36.0
python-dotenv>=1.0.1
